pyproject.toml
[tool.poetry]
name = "rasp0w"
version = "0.1.0"
description = ""
authors = ["rkjar <rkjar@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
